Hydrate Main with develop

+++ src/datalayer/syncService.js
@@ -3,26 +3,23 @@ import _ from 'lodash';
 import { decodeHex, decodeDataLayerResponse } from '../utils/datalayer-utils';
 import { Organization, Staging, ModelKeys } from '../models';
 import { getConfig } from '../utils/config-loader';
+import { logger } from '../config/logger.cjs';
 
 import * as dataLayer from './persistance';
 import * as simulator from './simulator';
-
-import Debug from 'debug';
-Debug.enable('climate-warehouse:datalayer:syncService');
-const log = Debug('climate-warehouse:datalayer:syncService');
 
 const { USE_SIMULATOR } = getConfig().APP;
 
 const POLLING_INTERVAL = 5000;
 const frames = ['-', '\\', '|', '/'];
 
-log('Start Datalayer Update Polling');
+logger.debug('Start Datalayer Update Polling');
 const startDataLayerUpdatePolling = async () => {
   const updateStoreInfo = await dataLayerWasUpdated();
   if (updateStoreInfo.length) {
     await Promise.all(
       updateStoreInfo.map(async (store) => {
-        log(
+        logger.debug(
           `Updates found syncing storeId: ${store.storeId} ${
             frames[Math.floor(Math.random() * 3)]
           }`,
@@ -74,21 +71,29 @@ const syncDataLayerStoreToClimateWarehouse = async (storeId, rootHash) => {
         storeData.keys_values.map(async (kv) => {
           const key = decodeHex(kv.key.replace(`${storeId}_`, ''));
           const modelKey = key.split('|')[0];
-          const value = JSON.parse(decodeHex(kv.value));
+          let value;
 
-          await ModelKeys[modelKey].create(value);
+          try {
+            value = JSON.parse(decodeHex(kv.value));
+          } catch (err) {
+            logger.error(`Cant parse json value: ${decodeHex(kv.value)}`);
+          }
 
-          const stagingUuid =
-            modelKey === 'unit'
-              ? value.warehouseUnitId
-              : modelKey === 'project'
-              ? value.warehouseProjectId
-              : undefined;
+          if (ModelKeys[modelKey]) {
+            await ModelKeys[modelKey].upsert(value);
 
-          if (stagingUuid) {
-            await Staging.destroy({
-              where: { uuid: stagingUuid },
-            });
+            const stagingUuid =
+              modelKey === 'unit'
+                ? value.warehouseUnitId
+                : modelKey === 'project'
+                ? value.warehouseProjectId
+                : undefined;
+
+            if (stagingUuid) {
+              await Staging.destroy({
+                where: { uuid: stagingUuid },
+              });
+            }
           }
         }),
       );
@@ -186,8 +191,8 @@ const getSubscribedStoreData = async (
   if (!alreadySubscribed) {
     const response = await subscribeToStoreOnDataLayer(storeId, ip, port);
     if (!response.success) {
-      log(`Retrying...`, retry + 1);
-      log('...');
+      logger.debug(`Retrying...`, retry + 1);
+      logger.debug('...');
       await new Promise((resolve) =>
         setTimeout(() => resolve(), timeoutInterval),
       );
@@ -198,8 +203,8 @@ const getSubscribedStoreData = async (
   if (!USE_SIMULATOR) {
     const storeExistAndIsConfirmed = await dataLayer.getRoot(storeId, true);
     if (!storeExistAndIsConfirmed) {
-      log(`Retrying...`, retry + 1);
-      log('...');
+      logger.debug(`Retrying...`, retry + 1);
+      logger.debug('...');
       await new Promise((resolve) =>
         setTimeout(() => resolve(), timeoutInterval),
       );
@@ -214,11 +219,11 @@ const getSubscribedStoreData = async (
     encodedData = await dataLayer.getStoreData(storeId);
   }
 
-  log(encodedData?.keys_values);
+  logger.debug(encodedData?.keys_values);
 
   if (_.isEmpty(encodedData?.keys_values)) {
-    log(`Retrying...`, retry + 1);
-    log('...');
+    logger.debug(`Retrying...`, retry + 1);
+    logger.debug('...');
     await new Promise((resolve) =>
       setTimeout(() => resolve(), timeoutInterval),
     );
@@ -259,6 +264,19 @@ const getStoreData = async (storeId, callback, onFail, retry = 0) => {
   }
 };
 
+const getCurrentStoreData = async (storeId) => {
+  if (USE_SIMULATOR) {
+    return [];
+  }
+
+  const encodedData = await dataLayer.getStoreData(storeId);
+  if (encodedData) {
+    return decodeDataLayerResponse(encodedData);
+  } else {
+    return [];
+  }
+};
+
 const getStoreIfUpdated = async (
   storeId,
   lastRootHash,
@@ -268,7 +286,7 @@ const getStoreIfUpdated = async (
 ) => {
   const rootResponse = await dataLayer.getRoot(storeId);
   if (rootResponse.confirmed && rootResponse.hash !== lastRootHash) {
-    log(`Updating orgUid ${storeId} with hash ${rootResponse.hash}`);
+    logger.debug(`Updating orgUid ${storeId} with hash ${rootResponse.hash}`);
     onUpdate(rootResponse.hash);
     await getStoreData(storeId, callback, onFail);
   }
@@ -285,4 +303,5 @@ export default {
   getStoreData,
   getStoreIfUpdated,
   POLLING_INTERVAL,
+  getCurrentStoreData,
 };
